requirements.txt
altgraph==0.17
astroid==2.4.2
auto-py-to-exe==2.7.2
beautifulsoup4==4.9.1
bottle==0.12.18
bottle-websocket==0.2.9
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
chromedriver==2.24.1
click==7.1.2
colorama==0.4.3
cPython==0.0.5
cycler==0.10.0
Cython==0.29.17
Django==2.2.14
django-crispy-forms==1.9.2
dnspython==1.16.0
Eel==0.11.0
emoji==0.5.4
et-xmlfile==1.0.1
future==0.18.2
gcloud==0.18.3
gevent==20.5.0
gevent-websocket==0.10.1
googleapis-common-protos==1.51.0
greenlet==0.4.15
gunicorn==20.0.4
html5lib==1.1
httplib2==0.17.3
idna==2.9
isort==4.3.21
jdcal==1.4.1
joblib==0.14.1
jws==0.1.3
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
lxml==4.5.2
matplotlib==3.0.3
mccabe==0.6.1
MouseInfo==0.1.3
numpy==1.18.5
oauth2client==4.1.2
openpyxl==2.6.4
pandas==0.25.3
pdfkit==0.6.1
pefile==2019.4.18
Pillow==7.1.2
protobuf==3.12.0rc1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.50
pycparser==2.20
pycryptodome==3.9.7
pyemojify==0.2.0
PyGetWindow==0.0.8
PyInstaller==3.6
pylint==2.5.3
pymongo==3.10.1
PyMsgBox==1.0.8
pyparsing==2.4.7
pyperclip==1.8.0
pypiwin32==223
Pyrebase4==4.3.0
PyRect==0.1.4
PyScreeze==0.1.26
python-dateutil==2.8.1
python-dotenv==0.14.0
python-jwt==2.0.1
PyTweening==1.0.3
pytz==2020.1
pywin32==227
pywin32-ctypes==0.2.0
requests==2.23.0
requests-toolbelt==0.9.1
rsa==4.0
scikit-learn==0.22.2.post1
scipy==1.4.1
selenium==3.141.0
six==1.14.0
socks==0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
UNKNOWN==0.0.0
urllib3==1.25.9
webencodings==0.5.1
whichcraft==0.6.1
whitenoise==5.2.0
wkhtmltopdf==0.2
wrapt==1.12.1
